+++ stream-compositions/services/transaction-composition-service/src/main/java/com/backbase/stream/compositions/transaction/core/mapper/DateOffsetMapper.java
@@ -1,4 +1,0 @@
-package com.backbase.stream.compositions.transaction.core.mapper;
-
-public class DateOffsetMapper {
-}